starting point
made a copy of the source hn form

copied the javascript code from w3schools that converts the form field
text to uppercase on the 'onblur' event. put this code in index.html and
tested it in chrome. it worked fine.

then put this code in the popup plugin and tested. it worked fine. in
the popup on the NH site, the text i had typed turned to uppercase when
i tabbed out.